added items on board
