+++ engine.py
@@ -67,108 +67,6 @@ def create_board(width, height): # Greg
     Returns:
     list: Game board
     '''
-   
-# jak połączyć to co zbieramy z planszy z wartościami w inventory? (add i remove?)
-
-# added_items = "kupa", "kupa", "miecz", "łuk"
-added_items = {"kupa": 13, "MANA": 3, "strzały": 5}
-removed_items = {"kupa": 12, "MANA": 1, "strzały": 5}
-
-def create_item(): # Greg
-    # to jest inwentarz przedmiotów które można zbierać na planszy (dopisać narzędzia bojowe?)
-    inventory = {}
-    inventory["MEDICNE"] = 1
-    inventory["FOOD"] = 1
-    inventory["DRINK"] = 1
-    inventory["MANA"] = 1
-    return inventory
-    
-def display_inventory(inventory):
-    for item in inventory:
-        print(f"{item}: {inventory[item]}")
-
-def add_to_inventory(inventory, added_items):
-    # tu chyba trzeba dopisać funkcję która będzie dodawała punkty do poszczególnych rzeczy w inwentarzu
-    for item in added_items:
-        if item in inventory:
-            inventory[item] += added_items[item]
-        else:
-            inventory[item] = added_items[item]
-    print(inventory)
-
-def remove_from_inventory(inventory, removed_items):
-    # tu trzeba dopisać jak poszczególne rzeczy będą ubywać w walce lub w trakcie np chodzenia
-    for item in removed_items:
-        if item in inventory:
-            inventory[item] -= removed_items[item]
-            if inventory[item] == 0:
-                del inventory[item]
-    return inventory
-
-def print_inventory_table(inventory, order):
-    print("""
------------------
-item name | count
------------------""")
-    order = order.replace(" ", "")
-    if order == "empty":
-        for item in inventory:
-            print(f"{item} | {inventory[item]}")
-    elif order == "count,asc":
-        sorted_dictionary = sorted(inventory.items(), key=select_value_of_key)
-        for item in sorted_dictionary:
-            print(f"{item[0]} | {item[1]}")
-    elif order == "count,desc":
-        sorted_dictionary = sorted(inventory.items(), key=select_value_of_key)
-        reversed_dictionary = sorted_dictionary[::-1]
-        for item in reversed_dictionary:
-            print(f"{item[0]} | {item[1]}")
-        
-def select_value_of_key(item):
-    return item[1]
-
-def import_inventory(inventory, filename = "inventory.csv"):
-    if os.path.isfile(filename):
-        with open(filename, "r") as file:
-            csvreader = csv.reader(file, skipinitialspace=True)
-            for row in csvreader:
-                for i in range(0, len(row)):
-                    if row[i] in inventory:
-                        inventory[row[i]] += 1
-                    else:
-                        inventory[row[i]] = 1
-        print(inventory)
-    else:
-        print(f"File '{filename}' not found!")
-            
-def export_inventory(inventory, filename = "inventory.csv"):
-    export_data = []
-    for item in inventory:
-        for i in range(0, inventory[item]):
-            export_data.append(item)
-    try:
-        with open(filename, "w", encoding="UTF8") as f:
-            writer = csv.writer(f)
-            writer.writerow(export_data)
-        print("Data export done!")
-    except:
-        print(f"You don't have permission creating file '{filename}'!")
-
-# testowanie funkcji importu i exportu, add i remove from inventory  
-# inventory = {}
-# import_inventory(inventory, "inventory.csv")
-# print_inventory_table(inventory, "count,desc")
-# inventory = create_item()
-# print_inventory_table(inventory, "count,desc")
-# add_to_inventory(inventory, added_items)
-# print_inventory_table(inventory, "count,desc")
-# export_inventory(inventory, "inventory.csv")
-# print_inventory_table(inventory, "count,desc")
-# remove_from_inventory(inventory, removed_items)
-# print_inventory_table(inventory, "count,desc")
-# export_inventory(inventory, "inventory.csv")
-# import_inventory(inventory, "inventory.csv")
-# print_inventory_table(inventory, "count,desc")
 
     
 def put_player_on_board(board, player):
