--- inventory.csv
@@ -1,1 +1,0 @@
-MEDICNE,FOOD,DRINK,MANA,MANA,MANA,kupa
